separate components from presonView
