+++ frontend/app/src/pages/Modals.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const Modals = () => {
+  return (
+    <div />
+  )
+}
