+++ frontend/app/src/pages/People/PersonPage.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import { useRouteMatch } from 'react-router-dom';
+
+export const PersonPage = () => {
+  const { path } = useRouteMatch();
+  return (
+    <div>{path}</div>
+  )
+}
