+++ frontend/app/src/pages/Tickets/index.tsx
@@ -1,0 +1,9 @@
+import { PeopleBody } from 'people/main'
+import React from 'react'
+import { Route } from 'react-router-dom'
+
+export const TicketsPage = () => {
+  return (
+    <PeopleBody selectedWidget="wanted" />
+  )
+}
