+++ frontend/app/src/people/PeopleList/index.ts
@@ -1,0 +1,1 @@
+export { PeopleList} from './PeopleList'
