+++ frontend/app/src/people/UserInfo/index.tsx
@@ -1,0 +1,13 @@
+import { useIsMobile } from 'hooks'
+import React from 'react'
+import { UserInfoDesktopView } from './UserInfoDesktopView'
+import { UserInfoMobileView } from './UserInfoMobileView'
+
+type UserInfoProps = 
+  { setShowQR, setShowSupport}
+
+export const UserInfo = (props: UserInfoProps) => {
+  const isMobile = useIsMobile()
+
+  return isMobile ? <UserInfoMobileView {...props}/> : <UserInfoDesktopView {...props} />
+}
