+++ frontend/app/src/people/personViewSlim.tsx
@@ -13,38 +13,36 @@ import TwitterView from './widgetViews/twitterView';
 import WantedView from './widgetViews/wantedView';
 
 import { useHistory, useLocation } from 'react-router';
+import { Button, IconButton, Modal } from '../components/common';
 import { meSchema } from '../components/form/schema';
-import { useIsMobile, usePageScroll } from '../hooks';
-import { Button, IconButton, Modal, SearchTextInput } from '../components/common';
-import { queryLimit } from '../store/main';
+import { useIsMobile } from '../hooks';
+import { PeopleList } from './PeopleList';
+import { UserInfo } from './UserInfo';
 import FocusedView from './main/focusView';
-import Person from './person';
+import { Widget } from './main/types';
 import Badges from './utils/badges';
 import ConnectCard from './utils/connectCard';
 import { widgetConfigs } from './utils/constants';
-import NoResults from './utils/noResults';
 import NoneSpace from './utils/noneSpace';
 import PageLoadSpinner from './utils/pageLoadSpinner';
 import { PostBounty } from './widgetViews/postBounty';
-import { Widget } from './main/types';
 
-const host = getHost();
-function makeQR(pubkey: string) {
-  return `sphinx.chat://?action=person&host=${host}&pubkey=${pubkey}`;
-}
-
-export default function PersonView(props: any) {
-  const { personId, loading, selectPerson, goBack } = props;
+export default function PersonView({loading = false}) {
   // on this screen, there will always be a pubkey in the url, no need for personId
-
   const { main, ui } = useStores();
   const { meInfo, peoplePageNumber } = ui || {};
+  const history = useHistory();
+  const location = useLocation();
+  
+  const personId = ui.selectedPerson;
+  function goBack() {
+    ui.setSelectingPerson(0);
+    history.goBack();
+  }
 
   const [loadingPerson, setLoadingPerson]: any = useState(false);
   const [loadedPerson, setLoadedPerson]: any = useState(null);
 
-  const history = useHistory();
-  const location = useLocation();
   const pathname = history?.location?.pathname;
   // FOR PEOPLE VIEW
   let person: any = main.people && main.people.length && main.people.find((f) => f.id === personId);
@@ -82,15 +80,8 @@ export default function PersonView(props: any) {
 
   const [showQR, setShowQR] = useState(false);
   const [showFocusView, setShowFocusView] = useState(false);
-  const qrString = makeQR(owner_pubkey || '');
 
-  async function loadMorePeople(direction) {
-    let newPage = peoplePageNumber + direction;
-    if (newPage < 1) {
-      newPage = 1;
-    }
-    await main.getPeople({ page: newPage });
-  }
+
 
   // if no people, load people on mount
   useEffect(() => {
@@ -164,11 +155,6 @@ export default function PersonView(props: any) {
     [updatePath]
   );
 
-  function selectPersonWithinFocusView(id, unique_name, pubkey) {
-    setShowFocusView(false);
-    setFocusIndex(-1);
-    selectPerson(id, unique_name, pubkey);
-  }
 
   useEffect(() => {
     if (ui.personViewOpenTab) {
@@ -184,10 +170,6 @@ export default function PersonView(props: any) {
     goBack();
   }
 
-  const { loadingTop, loadingBottom, handleScroll } = usePageScroll(
-    () => loadMorePeople(1),
-    () => loadMorePeople(-1)
-  );
 
   if (loading) return <div>Loading...</div>;
 
@@ -446,7 +428,6 @@ export default function PersonView(props: any) {
     }
   }
 
-  const defaultPic = '/static/person_placeholder.png';
   const mediumPic = img;
 
   function renderMobileView() {
@@ -513,42 +494,7 @@ export default function PersonView(props: any) {
               <div />
             )}
           </div>
-
-          {/* profile photo */}
-          <Head>
-            <Img src={mediumPic || defaultPic} />
-            <RowWrap>
-              <Name>{owner_alias}</Name>
-            </RowWrap>
-
-            {/* only see buttons on other people's profile */}
-            {canEdit ? (
-              <div style={{ height: 40 }} />
-            ) : (
-              <RowWrap style={{ marginBottom: 30, marginTop: 25 }}>
-                <a href={qrString}>
-                  <Button
-                    text="Connect"
-                    onClick={(e) => e.stopPropagation()}
-                    color="primary"
-                    height={42}
-                    width={120}
-                  />
-                </a>
-
-                <div style={{ width: 15 }} />
-
-                <Button
-                  text="Send Tip"
-                  color="link"
-                  height={42}
-                  width={120}
-                  onClick={() => setShowSupport(true)}
-                />
-              </RowWrap>
-            )}
-          </Head>
-
+          <UserInfo setShowQR={setShowQR} setShowSupport={setShowSupport} />
           <Tabs>
             {tabs &&
               Object.keys(tabs).map((name, i) => {
@@ -608,15 +554,6 @@ export default function PersonView(props: any) {
     );
   }
 
-  const loaderTop = <PageLoadSpinner show={loadingTop} />;
-  const loaderBottom = (
-    <PageLoadSpinner
-      noAnimate
-      show={loadingBottom}
-      style={{ position: 'absolute', bottom: 0, left: 0 }}
-    />
-  );
-
   function renderDesktopView() {
     const focusedDesktopModalStyles = newSelectedWidget
       ? {
@@ -633,184 +570,9 @@ export default function PersonView(props: any) {
         }}
       >
         {!canEdit && (
-          <PeopleList>
-            <DBack>
-              <Button color="clear" leadingIcon="arrow_back" text="Back" onClick={goBack} />
-
-              <SearchTextInput
-                small
-                name="search"
-                type="search"
-                placeholder="Search"
-                value={ui.searchText}
-                style={{
-                  width: 120,
-                  height: 40,
-                  border: '1px solid #DDE1E5',
-                  background: '#fff'
-                }}
-                onChange={(e) => {
-                  console.log('handleChange', e);
-                  ui.setSearchText(e);
-                }}
-              />
-            </DBack>
-
-            <PeopleScroller
-              style={{ width: '100%', overflowY: 'auto', height: '100%' }}
-              onScroll={handleScroll}
-            >
-              {loaderTop}
-              {people?.length ? (
-                people.map((t) => (
-                  <Person
-                    {...t}
-                    key={t.id}
-                    selected={personId === t.id}
-                    hideActions={true}
-                    small={true}
-                    select={selectPersonWithinFocusView}
-                  />
-                ))
-              ) : (
-                <NoResults />
-              )}
-
-              {/* make sure you can always scroll ever with too few people */}
-              {people?.length < queryLimit && <div style={{ height: 400 }} />}
-            </PeopleScroller>
-
-            {loaderBottom}
-          </PeopleList>
+          <PeopleList />
         )}
-
-        <AboutWrap
-          style={{
-            width: 364,
-            minWidth: 364,
-            background: '#ffffff',
-            color: '#000000',
-            padding: 40,
-            zIndex: 5,
-            // height: '100%',
-            marginTop: canEdit ? 64 : 0,
-            height: canEdit ? 'calc(100% - 64px)' : '100%',
-            borderLeft: '1px solid #ebedef',
-            borderRight: '1px solid #ebedef',
-            boxShadow: '1px 2px 6px -2px rgba(0, 0, 0, 0.07)'
-          }}
-        >
-          {canEdit && (
-            <div
-              style={{
-                position: 'absolute',
-                top: 0,
-                left: 0,
-                display: 'flex',
-                background: '#ffffff',
-                justifyContent: 'space-between',
-                alignItems: 'center',
-                width: 364,
-                minWidth: 364,
-                boxShadow: '0px 2px 6px rgba(0, 0, 0, 0.07)',
-                borderBottom: 'solid 1px #ebedef',
-                paddingRight: 10,
-                height: 64,
-                zIndex: 0
-              }}
-            >
-              <Button color="clear" leadingIcon="arrow_back" text="Back" onClick={goBack} />
-              <div />
-            </div>
-          )}
-
-          {/* profile photo */}
-          <Head>
-            <div style={{ height: 35 }} />
-
-            <Img src={mediumPic || defaultPic}>
-              <IconButton
-                iconStyle={{ color: '#5F6368' }}
-                style={{
-                  zIndex: 2,
-                  width: '40px',
-                  height: '40px',
-                  padding: 0,
-                  background: '#ffffff',
-                  border: '1px solid #D0D5D8',
-                  boxSizing: 'borderBox',
-                  borderRadius: 4
-                }}
-                icon={'qr_code_2'}
-                onClick={() => setShowQR(true)}
-              />
-            </Img>
-
-            <RowWrap>
-              <Name>{owner_alias}</Name>
-            </RowWrap>
-
-            {/* only see buttons on other people's profile */}
-            {canEdit ? (
-              <RowWrap
-                style={{
-                  marginBottom: 30,
-                  marginTop: 25,
-                  justifyContent: 'space-around'
-                }}
-              >
-                <Button
-                  text="Edit Profile"
-                  onClick={() => {
-                    switchWidgets('about');
-                    setShowFocusView(true);
-                  }}
-                  color="widget"
-                  height={42}
-                  style={{ fontSize: 13, background: '#f2f3f5' }}
-                  leadingIcon={'edit'}
-                  iconSize={15}
-                />
-                <Button
-                  text="Sign out"
-                  onClick={logout}
-                  height={42}
-                  style={{ fontSize: 13, color: '#3c3f41' }}
-                  iconStyle={{ color: '#8e969c' }}
-                  iconSize={15}
-                  color="white"
-                  leadingIcon="logout"
-                />
-              </RowWrap>
-            ) : (
-              <RowWrap
-                style={{
-                  marginBottom: 30,
-                  marginTop: 25,
-                  justifyContent: 'space-between'
-                }}
-              >
-                <Button
-                  text="Connect"
-                  onClick={() => setShowQR(true)}
-                  color="primary"
-                  height={42}
-                  width={120}
-                />
-
-                <Button
-                  text="Send Tip"
-                  color="link"
-                  height={42}
-                  width={120}
-                  onClick={() => setShowSupport(true)}
-                />
-              </RowWrap>
-            )}
-          </Head>
-
-          {renderWidgets('about')}
-        </AboutWrap>
+        <UserInfo setShowQR={setShowQR} setShowSupport={setShowSupport} />
 
         <div
           style={{
@@ -988,74 +750,6 @@ interface PanelProps {
   isMobile: boolean;
 }
 
-const PeopleList = styled.div`
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  background: #ffffff;
-  width: 265px;
-  overflow-y: overlay !important;
-
-  * {
-    scrollbar-width: 6px;
-    scrollbar-color: rgba(176, 183, 188, 0.25);
-  }
-
-  /* Works on Chrome, Edge, and Safari */
-  *::-webkit-scrollbar {
-    width: 6px;
-  }
-
-  *::-webkit-scrollbar-thumb {
-    background-color: rgba(176, 183, 188, 0.25);
-    background: rgba(176, 183, 188, 0.25);
-    width: 6px;
-    border-radius: 10px;
-    background-clip: padding-box;
-  }
-
-  ::-webkit-scrollbar-track-piece:start {
-    background: transparent url('images/backgrounds/scrollbar.png') repeat-y !important;
-  }
-
-  ::-webkit-scrollbar-track-piece:end {
-    background: transparent url('images/backgrounds/scrollbar.png') repeat-y !important;
-  }
-`;
-
-const PeopleScroller = styled.div`
-  overflow-y: overlay !important;
-  width: 100%;
-  height: 100%;
-`;
-
-const AboutWrap = styled.div`
-  overflow-y: auto !important;
-  ::-webkit-scrollbar-thumb {
-    background-color: rgba(176, 183, 188, 0);
-    background: rgba(176, 183, 188, 0);
-  }
-
-  &:hover {
-    ::-webkit-scrollbar-thumb {
-      background-color: rgba(176, 183, 188, 0.45);
-      background: rgba(176, 183, 188, 0.45);
-    }
-  }
-`;
-
-const DBack = styled.div`
-  min-height: 64px;
-  height: 64px;
-  display: flex;
-  padding-right: 10px;
-  align-items: center;
-  justify-content: space-between;
-  background: #ffffff;
-  box-shadow: 0px 1px 6px rgba(0, 0, 0, 0.07);
-  z-index: 0;
-`;
-
 const Panel = styled.div<PanelProps>`
   position: relative;
   background: #ffffff;
@@ -1120,51 +814,4 @@ const Tab = styled.div<TagProps>`
   cursor: pointer;
 `;
 
-const Head = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-`;
-
-const Name = styled.div`
-  font-style: normal;
-  font-weight: 500;
-  font-size: 24px;
-  line-height: 28px;
-  /* or 73% */
-
-  text-align: center;
-
-  /* Text 2 */
-
-  color: #3c3f41;
-`;
-
 const Sleeve = styled.div``;
-
-const RowWrap = styled.div`
-  display: flex;
-  justify-content: center;
-
-  width: 100%;
-`;
-
-interface ImageProps {
-  readonly src: string;
-}
-
-const Img = styled.div<ImageProps>`
-  background-image: url('${(p) => p.src}');
-  background-position: center;
-  background-size: cover;
-  margin-bottom: 20px;
-  width: 150px;
-  height: 150px;
-  border-radius: 50%;
-  position: relative;
-  display: flex;
-  align-items: flex-end;
-  justify-content: flex-end;
-`;
